% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ThermoPheno_functions.R
\name{default_parameters}
\alias{default_parameters}
\title{Default ThermoPheno crop parameters}
\usage{
default_parameters(crop_type = c("summer", "winter"))
}
\arguments{
\item{crop_type}{\code{summer} or \code{winter}.}
}
\value{
A named list of default crop parameters.
}
\description{
Default ThermoPheno crop parameters
}
